src/requirements.txt
python-dotenv
gunicorn
cloudinary
blinker
bson
click
colorama
decorator
dnspython
Flask
Flask-Cors
Flask-Login
Flask-SQLAlchemy
Flask-WTF
greenlet
itsdangerous
Jinja2
MarkupSafe
psycopg2-binary
pymongo
python-dateutil
self
six
SQLAlchemy
typing_extensions
Werkzeug
WTForms
